reorg to fit existing mod structure
